requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiounittest==1.3.0
async-timeout==3.0.1
asynctest==0.13.0
attrs==21.4.0
azure-core==1.23.1
azure-cosmos==3.2.0
azure-storage-blob==12.7.0
azure-storage-queue==12.1.5
Babel==2.7.0
botbuilder-azure==4.14.2
botbuilder-core==4.14.2
botbuilder-dialogs==4.14.2
botbuilder-integration-aiohttp==4.14.2
botbuilder-schema==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==3.4.8
datedelta==1.4
emoji==1.7.0
frozenlist==1.3.0
grapheme==0.6.0
idna==2.10
importlib-metadata==4.11.3
isodate==0.6.1
jsonpickle==1.4.2
msal==1.6.0
msrest==0.6.19
multidict==6.0.2
multipledispatch==0.6.0
oauthlib==3.2.0
pycparser==2.21
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
wincertstore==0.2
yarl==1.4.2
zipp==3.8.0
